Rework common_environment to be first-class python
Convert common_environment_py_content (string) into a first-class
python file common_environment_content.py and use the module's
__path__ to locate and read its contents.

This means one can edit common_environment_content with full syntax
highlighting and run tests against it, etc.
